requirements.txt
aws-cdk-lib==2.117.0
constructs>=10.0.0,<11.0.0
python-dotenv==1.0.1
boto3==1.34.29